fix(core): fix an edge case for split_off operation for RrbVec (#105)
There is an edge case that brakes the `push` operation on `RrbVec`. Here
are the steps to reproduce it:
- Create a new `RrbVec` instance exactly of the `BRANCH_FACTOR *
BRANCH_FACTOR` size.
- Split it exactly at the following index: `BRANCH_FACTOR *
BRANCH_FACTOR - BRANCH_FACTOR`.
- Try pushing a new item onto the new (right-hand side) vector.

A simple example for a vector that has a branching factor of 4:
- The in-memory representation of the following vector is going to be a
complete and full tree: `[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,
14, 15]`.
- Trying to split it at index 12 will produce two new vectors: `[0, 1,
2, 3, 4, 5, 6, 7, 8, 9, 10, 11]` and `[12, 13, 14, 15]`.

The right hand size vector will have a full tail. The bug was that the
tail was not pushed onto tree when full. When trying to push an ordinary
item, `16` for example, it will overflow the vector and cause a panic.